refactor(menu): Move template menu list out to a separate file (#634)
